feat: update services to use factory pattern
