--- server/src/service/emailService.js
@@ -1,67 +1,0 @@
-import nodemailer from "nodemailer";
-import config from "../config/config.js";
-import logger from "../config/logger.js";
-
-const transport = nodemailer.createTransport(config.email.smtp);
-/* istanbul ignore next */
-if (config.env !== "test") {
-  transport
-    .verify()
-    .then(() => logger.info("Connected to email server"))
-    .catch(() =>
-      logger.warn(
-        "Unable to connect to email server. Make sure you have configured the SMTP options in .env"
-      )
-    );
-}
-
-/**
- * Send an email
- * @param {string} to
- * @param {string} subject
- * @param {string} text
- * @returns {Promise}
- */
-const sendEmail = async (to, subject, text) => {
-  const msg = { from: config.email.from, to, subject, text };
-  await transport.sendMail(msg);
-};
-
-/**
- * Send reset password email
- * @param {string} to
- * @param {string} token
- * @returns {Promise}
- */
-const sendResetPasswordEmail = async (to, token) => {
-  const subject = "Reset password";
-  // replace this url with the link to the reset password page of your front-end app
-  const resetPasswordUrl = `${config.siteUrls.serverURL}/reset-password?token=${token}`;
-  const text = `Dear user,
-To reset your password, click on this link: ${resetPasswordUrl}
-If you did not request any password resets, then ignore this email.`;
-  await sendEmail(to, subject, text);
-};
-
-/**
- * Send verification email
- * @param {string} to
- * @param {string} token
- * @returns {Promise}
- */
-const sendVerificationEmail = async (to, token) => {
-  const subject = "Email Verification";
-  // replace this url with the link to the email verification page of your front-end app
-  const verificationEmailUrl = `${config.siteUrls.serverURL}/verify-email?token=${token}`;
-  const text = `Dear user,
-To verify your email, click on this link: ${verificationEmailUrl}
-If you did not create an account, then ignore this email.`;
-  await sendEmail(to, subject, text);
-};
-
-export default {
-  transport,
-  sendEmail,
-  sendResetPasswordEmail,
-  sendVerificationEmail
-};
