--- server/src/service/tokenService.js
@@ -1,151 +1,0 @@
-import jwt from "jsonwebtoken";
-import moment from "moment-timezone";
-import config from "../config/config.js";
-import userService from "./userService.js";
-import Token from "../models/token.js";
-import ApiError from "../utils/ApiError.js";
-// eslint-disable-next-line
-import enums from "../json/enums.json" assert { type: "json" };
-
-/**
- * Generate token
- * @param {ObjectId} userId
- * @param {Moment} expires
- * @param {string} type
- * @param {string} [secret]
- * @returns {string}
- */
-const generateToken = (userId, expires, type, secret = config.jwt.secret) => {
-  const payload = {
-    sub: userId,
-    iat: moment().unix(),
-    exp: expires.unix(),
-    type
-  };
-  return jwt.sign(payload, secret);
-};
-
-/**
- * Save a token
- * @param {string} token
- * @param {ObjectId} userId
- * @param {Moment} expires
- * @param {string} type
- * @param {boolean} [blacklisted]
- * @returns {Promise<Token>}
- */
-const saveToken = async (token, userId, expires, type, blacklisted = false) => {
-  const tokenDoc = await Token.create({
-    token,
-    user: userId,
-    expires: expires.toDate(),
-    type,
-    blacklisted
-  });
-  return tokenDoc;
-};
-
-/**
- * Verify token and return token doc (or throw an error if it is not valid)
- * @param {string} token
- * @param {string} type
- * @returns {Promise<Token>}
- */
-const verifyToken = async (token, type) => {
-  const payload = jwt.verify(token, config.jwt.secret);
-  const tokenDoc = await Token.findOne({
-    token,
-    type,
-    user: payload.sub,
-    blacklisted: false
-  });
-  if (!tokenDoc) {
-    throw new Error("Token not found");
-  }
-  return tokenDoc;
-};
-
-/**
- * Generate auth tokens
- * @param {User} user
- * @returns {Promise<Object>}
- */
-const generateAuthTokens = async (user) => {
-  const accessTokenExpires = moment().add(
-    config.jwt.accessExpirationMinutes,
-    "minutes"
-  );
-  const accessToken = generateToken(user.id, accessTokenExpires, "accessToken");
-
-  const refreshTokenExpires = moment().add(
-    config.jwt.refreshExpirationDays,
-    "days"
-  );
-  const refreshToken = generateToken(
-    user.id,
-    refreshTokenExpires,
-    "refreshToken"
-  );
-  await saveToken(refreshToken, user.id, refreshTokenExpires, "refreshToken");
-
-  return {
-    access: {
-      token: accessToken,
-      expires: accessTokenExpires.toDate()
-    },
-    refresh: {
-      token: refreshToken,
-      expires: refreshTokenExpires.toDate()
-    }
-  };
-};
-
-/**
- * Generate reset password token
- * @param {string} email
- * @returns {Promise<string>}
- */
-const generateResetPasswordToken = async (email) => {
-  const user = await userService.getUserByEmail(email);
-  if (!user) {
-    throw new ApiError(
-      enums.HTTP_CODES.NOT_FOUND,
-      "No users found with this email"
-    );
-  }
-  const expires = moment().add(
-    config.jwt.resetPasswordExpirationMinutes,
-    "minutes"
-  );
-  const resetPasswordToken = generateToken(
-    user.id,
-    expires,
-    "resetPasswordToken"
-  );
-  await saveToken(resetPasswordToken, user.id, expires, "resetPasswordToken");
-  return resetPasswordToken;
-};
-
-/**
- * Generate verify email token
- * @param {User} user
- * @returns {Promise<string>}
- */
-const generateVerifyEmailToken = async (user) => {
-  const expires = moment().add(
-    config.jwt.verifyEmailExpirationMinutes,
-    "minutes"
-  );
-  const verifyEmailToken = generateToken(user.id, expires, "verifyEmailToken");
-  await saveToken(verifyEmailToken, user.id, expires, "verifyEmailToken");
-  return verifyEmailToken;
-};
-
-export default {
-  generateToken,
-  saveToken,
-  verifyToken,
-  generateAuthTokens,
-  generateResetPasswordToken,
-  generateVerifyEmailToken
-};
